Fix type conversion warnings.

+++ libyara/re.c
@@ -1061,7 +1061,7 @@ static int _yr_re_emit(
       if (bookmark_2 - bookmark_3 < INT32_MIN)
         return ERROR_REGULAR_EXPRESSION_TOO_LARGE;
 
-      repeat_args.offset = (int32_t)(bookmark_2 - bookmark_3);
+      repeat_args.offset = (int32_t) (bookmark_2 - bookmark_3);
 
       FAIL_ON_ERROR(_yr_emit_inst_arg_struct(
           emit_context,
@@ -1080,7 +1080,7 @@ static int _yr_re_emit(
       if (bookmark_4 - bookmark_1 > INT32_MAX)
         return ERROR_REGULAR_EXPRESSION_TOO_LARGE;
 
-      repeat_start_args_addr->offset = (int32_t)(bookmark_4 - bookmark_1);
+      repeat_start_args_addr->offset = (int32_t) (bookmark_4 - bookmark_1);
     }
 
     if (emit_split)
@@ -1114,7 +1114,7 @@ static int _yr_re_emit(
       split_offset_addr = (int16_t*) yr_arena_ref_to_ptr(
           emit_context->arena, &split_offset_ref);
 
-      *split_offset_addr = (int16_t)(bookmark_2 - bookmark_1);
+      *split_offset_addr = (int16_t) (bookmark_2 - bookmark_1);
     }
 
     break;
@@ -2139,8 +2139,8 @@ int yr_re_fast_exec(
       }
 
       bytes_matched = flags & RE_FLAGS_BACKWARDS
-                          ? input_data - current->input - 1
-                          : current->input - input_data;
+                          ? (int) (input_data - current->input - 1)
+                          : (int) (current->input - input_data);
       uint8_t mask;
       uint8_t value;
